Complete UI refinement and production configuration for Railway deployment (Sanitized)

+++ database/migrations/2026_01_24_130000_add_flashcard_id_to_messages_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('messages', function (Blueprint $table) {
-            $table->unsignedBigInteger('flashcard_id')->nullable()->after('content');
+            $table->unsignedInteger('flashcard_id')->nullable()->after('content');
             $table->foreign('flashcard_id')->references('id')->on('flashcard')->onDelete('set null');
         });
     }
